some fixups
