+++ platform-core-module/src/main/java/com/softicar/platform/core/module/module/IModule.java
@@ -5,8 +5,7 @@ import com.softicar.platform.core.module.module.instance.IModuleInstanceTable;
 import com.softicar.platform.emf.module.IEmfModule;
 
 /**
- * Interface of modules that use instances of type
- * {@link IModuleInstance}.
+ * Interface of modules that use instances of type {@link IModuleInstance}.
  *
  * @author Alexander Schmidt
  */
@@ -15,7 +14,7 @@ public interface IModule<I extends IModuleInstance<I>> extends IEmfModule<I> {
 	/**
 	 * Returns the table that represents the module instances.
 	 *
-	 * @return the module instance table (never null)
+	 * @return the module instance table (never <i>null</i>)
 	 */
 	IModuleInstanceTable<I> getModuleInstanceTable();
 }
