+++ platform-core-module/src/test/java/com/softicar/platform/core/module/permission/AbstractModulePermissionUiTest.java
@@ -3,8 +3,8 @@ package com.softicar.platform.core.module.permission;
 import com.softicar.platform.core.module.access.module.AbstractModuleTest;
 import com.softicar.platform.core.module.module.instance.AGModuleInstance;
 import com.softicar.platform.core.module.permission.assignment.AGModuleInstancePermissionAssignment;
-import com.softicar.platform.core.module.test.module.standard.alpha.TestModuleAlpha;
-import com.softicar.platform.core.module.test.module.standard.beta.TestModuleBeta;
+import com.softicar.platform.core.module.test.module.alpha.TestModuleAlpha;
+import com.softicar.platform.core.module.test.module.beta.TestModuleBeta;
 import com.softicar.platform.core.module.user.AGUser;
 import com.softicar.platform.core.module.uuid.AGUuid;
 import com.softicar.platform.emf.module.IEmfModule;
