+++ platform-core-module/src/test/java/com/softicar/platform/core/module/test/module/alpha/TestModuleAlphaInstance.java
@@ -1,4 +1,4 @@
-package com.softicar.platform.core.module.test.module.standard.alpha;
+package com.softicar.platform.core.module.test.module.alpha;
 
 import com.softicar.platform.core.module.module.instance.AGModuleInstance;
 import com.softicar.platform.core.module.module.instance.IModuleInstance;
