+++ platform-core-module/src/test/java/com/softicar/platform/core/module/test/module/beta/TestModuleBetaInstance.java
@@ -1,4 +1,4 @@
-package com.softicar.platform.core.module.test.module.standard.beta;
+package com.softicar.platform.core.module.test.module.beta;
 
 import com.softicar.platform.core.module.module.instance.AGModuleInstance;
 import com.softicar.platform.core.module.module.instance.IModuleInstance;
