+++ platform-core-module/src/test/java/com/softicar/platform/core/module/test/module/beta/TestModuleBetaPermissions.java
@@ -1,4 +1,4 @@
-package com.softicar.platform.core.module.test.module.standard.beta;
+package com.softicar.platform.core.module.test.module.beta;
 
 import com.softicar.platform.core.module.module.permission.TestModulePermission;
 import com.softicar.platform.emf.module.permission.IEmfModulePermission;
